requirements.txt
Django>=4.2
pymongo
dnspython
google-genai
gunicorn
python-dotenv
whitenoise